add SocketManager

+++ library-client-rtsp/src/main/java/com/alexvas/rtsp/GazeDataListener.kt
@@ -2,5 +2,5 @@ package com.alexvas.rtsp
 
 interface GazeDataListener {
 
-    public fun onGazeDataReady(gazeData: DoubleArray);
+    public fun onGazeDataReady(gazeData: FloatArray);
 }
